pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "whirltube"
version = "0.4.5"
description = "WhirlTube: lightweight GTK4 frontend for YouTube using mpv and yt-dlp"
readme = "README.md"
license = { text = "GPL-3.0-or-later" }
authors = [{ name = "WhirlTube contributors" }]
requires-python = ">=3.12"
dependencies = [
  "httpx>=0.27",
  "yt-dlp>=2024.8.6",
  "keyring>=24.0",
  "secretstorage>=3.3; sys_platform=='linux'",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.9",
  "mypy>=1.11",
  "pytest>=8.3",
  "pytest-cov>=5.0",
]
embed = [
  "python-mpv>=1.0.6",
]

[project.scripts]
whirltube = "whirltube.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/whirltube"]
include = [
  "src/whirltube/assets/**",
  "data/**",
]

[tool.hatch.build.targets.sdist]
include = [
  "src/**",
  "README.md",
  "CHANGELOG.md",
  "LICENSE",
  "pyproject.toml",
  "mypy.ini",
  "ruff.toml",
  "tests/**",
  "data/**",
  "flatpak/**",
]

[tool.ruff]
line-length = 100
target-version = "py312"
lint.select = ["E","F","W","I","UP","B","SIM","C90"]
lint.ignore = ["E203","E501"]
src = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
